bug fix: check already intalled

+++ recipes/default.rb
@@ -6,6 +6,7 @@ when 'centos'
   execute "install blast" do
     command "yum -y install ftp://ftp.ncbi.nlm.nih.gov/blast/executables/LATEST/ncbi-blast-2.2.29+-1.x86_64.rpm"
     action :run
+    not_if { File.exists?("/usr/bin/blastp") }
   end
 end
